# Conflicts:
#	python/pyproject.toml
#	python/src/yggdrasil/polars/cast.py
#	python/src/yggdrasil/types/cast/cast_options.py
#	python/src/yggdrasil/version.py

+++ python/tests/yggdrasil/databricks/sql/test_engine.py
@@ -48,11 +48,7 @@ class TestSQLEngine(unittest.TestCase):
 
         self.assertEqual(read.num_rows, data.num_rows + other_data.num_rows)
 
-        read = result.to_polars_lazy().collect()
-
-        self.assertEqual(read.shape[0], data.num_rows + other_data.num_rows)
-
-        read = result.to_polars()
+        read = result.to_polars(stream=False)
 
         self.assertEqual(read.shape[0], data.num_rows + other_data.num_rows)
 
@@ -82,24 +78,13 @@ class TestSQLEngine(unittest.TestCase):
 
     def test_warehouse_crud(self):
         warehouses = self.workspace.warehouses()
-        warehouse = warehouses.create(name="tmp warehouse")
 
-        self.assertEqual(warehouse.warehouse_name, "tmp warehouse")
+        warehouse = None
 
-        warehouse.delete()
+        try:
+            warehouse = warehouses.create(name="tmp warehouse", wait=False)
 
-    def test_sander(self):
-        from yggdrasil.databricks import Workspace
-
-        workspace = Workspace(
-            host="dbc-e6e40d20-0e8f.cloud.databricks.com"
-        )
-        engine = workspace.sql()
-
-        result = engine.execute("SELECT * FROM `trading`.ba_3mv_polaris__p__volcano_ref_input.curve_data limit 1000")
-
-        read = result.to_arrow_table()
-        read = result.to_pandas()
-        read = result.to_polars(stream=False)
-
-        print(read)
+            self.assertEqual(warehouse.warehouse_name, "tmp warehouse")
+        finally:
+            if warehouse:
+                warehouse.delete()
